handle alerts of duplicate names and remove tasks

+++ src/pages/Home.tsx
@@ -15,14 +15,27 @@ export function Home() {
   const [tasks, setTasks] = useState<Task[]>([]);
 
   function handleAddTask(newTaskTitle: string) {
+    const findTask = tasks.find(item => newTaskTitle === item.title);
+
+    if(!!findTask) {
+      Alert.alert(
+        'Task já cadastrada', 
+        'Você não pode cadastrar uma task com o mesmo nome'
+      );
+
+      return;
+    }
+      
     if (newTaskTitle !== '') {
       const data = {
         id: new Date().getTime(),
         title: newTaskTitle,
         done: false
       };
+
       return setTasks([...tasks, data]);
     }
+
     return Alert.alert('Favor adicionar uma tarefa!');
   }
 
@@ -39,24 +52,42 @@ export function Home() {
   }
 
   function handleRemoveTask(id: number) {
-    const deleteItem = tasks.filter(item => {
-      return item.id !== id 
-    })
-
-    setTasks(deleteItem);
+    Alert.alert(
+      'Remover item', 
+      'Tem certeza que você deseja remover esse item?', 
+      [
+        {
+          text: 'Não',
+        },
+        {
+          text: 'Sim',
+          onPress: () => {
+            const deleteItem = tasks.filter(item => {
+              return item.id !== id 
+            });
+            setTasks(deleteItem);
+          }
+        }
+      ], 
+      {
+        cancelable: true
+      } 
+    );
   }
 
   return (
     <>
       <Header />
+      <View style={{backgroundColor: '#B2B2B2', flex: 1}}>
+        <TodoInput addTask={handleAddTask} />
+        <MyTasksList 
+          tasks={tasks} 
+          onPress={handleMarkTaskAsDone} 
+          onLongPress={handleRemoveTask} 
+        />
+      </View>
 
-      <TodoInput addTask={handleAddTask} />
 
-      <MyTasksList 
-        tasks={tasks} 
-        onPress={handleMarkTaskAsDone} 
-        onLongPress={handleRemoveTask} 
-      />
     </>
   )
 }
